requirements.txt
python-dotenv
pydantic
typer
transformers
torch
numpy
pandas
pytest
pydantic-settings
rapidfuzz
